requirements.txt
librosa==0.10.1
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow>=2.13
keras
torch
torchaudio
transformers>=4.35
datasets
huggingface_hub
timm
tqdm
pyyaml
gradio
xgboost
soundfile
audioread
